refactor: split trace experiment support
