+++ tests/core/rig/spec_test.rs
@@ -3,7 +3,7 @@
 
 use crate::rig::{
     PipelineStep, RigResourcesSpec, RigSpec, ServiceKind, ServiceSpec, SharedPathSpec, SymlinkSpec,
-    TraceExperimentArtifactSpec, TraceVariantSpec, WorkloadEntry, WorkloadSpec,
+    TraceVariantSpec, WorkloadEntry, WorkloadSpec,
 };
 
 /// Canonical fixture matching the studio-playground-dev shape used as the
@@ -157,73 +157,6 @@ fn test_spec_trace_variants_parse_multi_component_overlays() {
     assert_eq!(variant.overlays[0].overlay, "overlays/studio.patch");
     assert_eq!(variant.overlays[1].component, "wordpress-playground");
     assert_eq!(variant.overlays[1].overlay, "overlays/playground.patch");
-}
-
-#[test]
-fn test_spec_trace_experiments_parse_lifecycle_settings_and_artifacts() {
-    let json = r#"{
-        "id": "studio-playground-dev",
-        "trace_experiments": {
-            "template-site": {
-                "setup": [
-                    { "command": "node bench/create-template-site.mjs", "cwd": "${package.root}" }
-                ],
-                "settings": {
-                    "STUDIO_TRACE_SITE_TEMPLATE": "/tmp/studio-template-site",
-                    "USE_TEMPLATE": true
-                },
-                "env": {
-                    "STUDIO_EXPERIMENT_MODE": "template"
-                },
-                "artifacts": [
-                    "/tmp/studio-template-site/report.json",
-                    { "label": "template log", "path": "/tmp/studio-template-site/template.log" }
-                ],
-                "teardown": [
-                    { "command": "rm -rf /tmp/studio-template-site" }
-                ]
-            }
-        }
-    }"#;
-    let spec: RigSpec = serde_json::from_str(json).expect("parse");
-    let experiment = spec
-        .trace_experiments
-        .get("template-site")
-        .expect("experiment");
-
-    assert_eq!(
-        experiment.setup[0].command,
-        "node bench/create-template-site.mjs"
-    );
-    assert_eq!(experiment.setup[0].cwd.as_deref(), Some("${package.root}"));
-    assert_eq!(
-        experiment.settings["STUDIO_TRACE_SITE_TEMPLATE"],
-        serde_json::Value::String("/tmp/studio-template-site".to_string())
-    );
-    assert_eq!(
-        experiment.settings["USE_TEMPLATE"],
-        serde_json::Value::Bool(true)
-    );
-    assert_eq!(
-        experiment
-            .env
-            .get("STUDIO_EXPERIMENT_MODE")
-            .map(String::as_str),
-        Some("template")
-    );
-    assert!(matches!(
-        &experiment.artifacts[0],
-        TraceExperimentArtifactSpec::Path(path) if path == "/tmp/studio-template-site/report.json"
-    ));
-    assert!(matches!(
-        &experiment.artifacts[1],
-        TraceExperimentArtifactSpec::Detailed { label, path }
-            if label == "template log" && path == "/tmp/studio-template-site/template.log"
-    ));
-    assert_eq!(
-        experiment.teardown[0].command,
-        "rm -rf /tmp/studio-template-site"
-    );
 }
 
 #[test]
